Moved propagation related modules to common namespace. Removed redundant classes
