+++ GS_example.py
@@ -5,9 +5,7 @@ Created on Thu Aug  4 15:44:04 2022
 @author: PK
 """
 
-from light_prop.propagation_params import PropagationParams
-from light_prop.propagation import ConvolutionPropagation, NNPropagation
-from light_prop.propagation_input import PropagationInput
+from light_prop.propagation.params import PropagationParams
 from light_prop.lightfield import LightField
 import light_prop.calculations as calc
 import numpy as np
@@ -33,14 +31,12 @@ phase = np.array(
       np.arange(-params.matrix_size / 2, params.matrix_size / 2) * params.pixel] for y in
      np.arange(-params.matrix_size / 2, params.matrix_size / 2) * params.pixel])
 
-
-
 GS = GerchbergSaxton(params)
 
 res = GS.optimize(LightField(amp, phase), LightField(target, phase), 5)
 
 plotter = GeneratePropagationPlot(res[1])
-plotter.save_output_as_figure("outs/structure.png", output_type = GeneratePropagationPlot.PLOT_PHASE)
+plotter.save_output_as_figure("outs/structure.png", output_type=GeneratePropagationPlot.PLOT_PHASE)
 
 plotter2 = GeneratePropagationPlot(res[0])
-plotter2.save_output_as_figure("outs/result.png", output_type = GeneratePropagationPlot.PLOT_ABS)
+plotter2.save_output_as_figure("outs/result.png", output_type=GeneratePropagationPlot.PLOT_ABS)
