+++ light_prop/propagation/methods.py
@@ -9,15 +9,13 @@ import numpy as np
 import logging
 from scipy import signal
 from tensorflow import keras
-from keras.models import Sequential
 from keras.layers import Convolution2D
-from keras.models import backend as K
-import tensorflow as tf
 
 from light_prop.calculations import h
-from light_prop.propagation_params import PropagationParams
-from light_prop.propagation_results import PropagationResult
+from light_prop.propagation.keras_layers import Aexp, ReIm_convert, Structure
+from light_prop.propagation.params import PropagationParams
 from light_prop.lightfield import LightField
+
 
 class BasePropagation:
     def __int__(self, propagation_params: PropagationParams):
@@ -56,45 +54,6 @@ class ConvolutionPropagation(BasePropagation):
 
     def calculate_propagation(self, field_distribution, field_modifier):
         return signal.fftconvolve(field_distribution, field_modifier, mode='same')
-
-
-
-class Aexp(keras.layers.Layer):
-    def __init__(self, **kwargs):
-        super(Aexp, self).__init__(**kwargs)
-
-    def call(self, inputs):
-        self.A = K.sqrt(K.square(inputs[:, 0]) + K.square(inputs[:, 1]))
-        self.phi = tf.math.atan2(inputs[:, 1], inputs[:, 0])
-        return K.concatenate([self.A, self.phi], axis=1)
-
-
-class ReIm_convert(keras.layers.Layer):
-    def __init__(self, **kwargs):
-        super(ReIm_convert, self).__init__(**kwargs)
-
-    def call(self, inputs):
-        self.Re = inputs[:, 0] * K.cos(inputs[:, 1])
-        self.Im = inputs[:, 0] * K.sin(inputs[:, 1])
-
-        return K.concatenate([self.Re, self.Im], axis=1)
-
-
-class Structure(keras.layers.Layer):
-    def __init__(self, kernel_initializer, **kwargs):
-        super(Structure, self).__init__(**kwargs)
-        self.kernel_initializer = kernel_initializer
-
-    def build(self, input_shape):
-        # Create a trainable weight variable for this layer.
-        self.kernel = self.add_weight(name='kernel',
-                                      shape=(input_shape[2], input_shape[3]),
-                                      initializer=self.kernel_initializer,  # TODO: Choose your initializer
-                                      trainable=True)
-        super(Structure, self).build(input_shape)
-
-    def call(self, inputs):
-        return K.concatenate([inputs[:, 0], inputs[:, 1] + self.kernel], axis=1)
 
 
 class NNPropagation(ConvolutionPropagation):
